Split main game code from launcher
