+++ FoFiX.py
@@ -26,13 +26,22 @@
 '''
 Main game executable.
 '''
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 
 import argparse
-import sys
-import os
-import platform
-import subprocess
 import atexit
+import os
+import subprocess
+import sys
+import traceback
+
+from fretwork import log
+
+from fofix.core import VFS, Version
+from fofix.core.Language import _
+
 
 def run_command(command):
     command = command.split(' ')
@@ -41,28 +50,9 @@ def run_command(command):
         stderr=subprocess.PIPE,
         stdin=subprocess.PIPE)
 
-    output, err = cmd.communicate()
-    data = output.strip()
+    output, __ = cmd.communicate()
+    return output.strip()
 
-    return data
-
-pisSet = None
-# This prevents the following message being displayed on osx:
-# ApplePersistenceIgnoreState: Existing state will not be touched. New state will be written to *path*
-if 'Darwin' in platform.platform():
-    data = run_command('defaults read org.python.python ApplePersistenceIgnoreState')
-
-    if data in ['1', 'ON']:
-        run_command('defaults write org.python.python ApplePersistenceIgnoreState 0')
-        atexit.register(run_command, 'defaults write org.python.python ApplePersistenceIgnoreState %s' % data)
-        apisSet = True
-    else:
-        apisSet = False
-
-# Add the directory of DLL dependencies to the PATH if we're running
-# from source on Windows so we pick them up when those bits are imported.
-if os.name == 'nt' and not hasattr(sys, 'frozen'):
-    os.environ['PATH'] = os.path.abspath(os.path.join('.', 'win32', 'deps', 'bin')) + os.pathsep + os.environ['PATH']
 
 def disable_gl_checks():
     assert 'OpenGL' not in sys.modules
@@ -76,239 +66,92 @@ def disable_gl_checks():
         OpenGL.SIZE_1_ARRAY_UNPACK = False
         OpenGL.STORE_POINTERS = False
 
+
 def cmd_args():
     parser = argparse.ArgumentParser(description='Frets On Fire X (FoFiX)')
-    # Where the name automatically assigned to the metavar option is to long i used x in its place.
-    # Might go back and
+
     options = parser.add_argument_group('Options')
-    options.add_argument('-r', '--resolution', type=str,  metavar='x',    help='Force a specific resolution to be used.', default=None)
-    options.add_argument('-f', '--fullscreen',  action='store_true',       help='Force usage of full-screen mode.', default=None)
-    options.add_argument('-c', '--config',     type=str,  metavar='x',     help='Use this configuration file instead of the fofix.ini in its default location.  Use "reset" to use the usual fofix.ini but clear it first.')
-    options.add_argument('-t', '--theme',      type=str,  metavar='x', help='Force the specified theme to be used. Remember to quote the theme name if it contains spaces (e.g. %(prog)s -t "Guitar Hero III")', default=None)
-    options.add_argument('-s', '--song',       type=str,                  help='Play a song in one-shot mode. (See "One-shot mode options" below.)')
+    options.add_argument('-r', '--resolution', type=str, metavar='x', help='Force a specific resolution to be used.', default=None)
+    options.add_argument('-f', '--fullscreen', action='store_true',   help='Force usage of full-screen mode.', default=None)
+    options.add_argument('-c', '--config',     type=str, metavar='x', help='Use this configuration file instead of the fofix.ini in its default location.  Use "reset" to use the usual fofix.ini but clear it first.')
+    options.add_argument('-t', '--theme',      type=str, metavar='x', help='Force the specified theme to be used. Remember to quote the theme name if it contains spaces (e.g. %(prog)s -t "Guitar Hero III").', default=None)
+    options.add_argument('-s', '--song',       type=str,              help='Play a song in one-shot mode (see "One-shot mode options" below).')
 
     osm = parser.add_argument_group('One-Shot Mode')
-    osm.add_argument('-p', '--part',    type=int, help='0: Guitar, 1: Rhythm, 2: Bass, 3: Lead')
-    osm.add_argument('-d', '--diff',    type=int, help='0: Expert, 1: Hard, 2: Medium, 3: Easy (Only applies if "part" is set)')
-    osm.add_argument('-m', '--mode',    type=int, help='0: Quickplay, 1: Practice, 2: Career')
+    osm.add_argument('-p', '--part',    type=int, help='0: Guitar, 1: Rhythm, 2: Bass, 3: Lead.')
+    osm.add_argument('-d', '--diff',    type=int, help='0: Expert, 1: Hard, 2: Medium, 3: Easy (Only applies if "part" is set).')
+    osm.add_argument('-m', '--mode',    type=int, help='0: Quickplay, 1: Practice, 2: Career.')
     osm.add_argument('-n', '--players', type=int, help='Number of multiplayer players.')
 
     adv = parser.add_argument_group('Advanced')
-    adv.add_argument('-v', '--verbose',        action='store_true', help='Verbose messages')
+    adv.add_argument('-v', '--verbose',        action='store_true', help='Verbose messages.')
     adv.add_argument('-g', '--gl-error-check', action='store_true', help='Enable OpenGL error checking.')
 
     return vars(parser.parse_args())
 
-args = cmd_args()
-
-#stump: disable pyOpenGL error checking if we are not asked for it.
-# This must be before *anything* that may import pyOpenGL!
-if not args['gl_error_check']:
-    disable_gl_checks()
-
-import traceback
-
-import pygame
-
-import fretwork
-
-from fretwork import log
-
-from fofix.core import Version
-from fofix.core import VFS
-
-# setup the logfile
-# File object representing the logfile.
-if os.name == "posix": # evilynux - logfile in ~/.fofix/ for GNU/Linux and MacOS X
-    # evilynux - Under MacOS X, put the logs in ~/Library/Logs
-    if os.uname()[0] == "Darwin":
-        logFile = open(os.path.expanduser('~/Library/Logs/%s.log' % Version.PROGRAM_UNIXSTYLE_NAME), 'w')
-    else: # GNU/Linux et al.
-        logFile = VFS.open('/userdata/%s.log' % Version.PROGRAM_UNIXSTYLE_NAME, 'w')
-else:
-    logFile = VFS.open('/userdata/%s.log' % Version.PROGRAM_UNIXSTYLE_NAME, 'w')
-
-log.setLogfile(logFile)
-
-import fretwork
-fretworkRequired = (0, 2, 0)
-reqVerStr = '.'.join([str(i) for i in fretworkRequired])
-fretworkErrorStr = '''
-
-The version of fretwork installed is old. Please install the latest version from github.
-https://github.com/fofix/fretwork/releases/
-Installed: {0}
-Required: {1}
-'''
-if hasattr(fretwork, '__version__'): # The first version of fretwork didnt have __version__
-    version = fretwork.__version__.split('-')[0] # remove 'dev' from ver
-    version = tuple([int(i) for i in version.split('.')])
-
-    if version < fretworkRequired:
-        fwErrStr = fretworkErrorStr.format(fretwork.__version__, reqVerStr)
-        raise RuntimeError(fwErrStr)
-
-else:
-    version = '0.1.0'
-    fwErrStr = fretworkErrorStr.format(version, reqVerStr)
-    raise RuntimeError(fwErrStr)
-
-from fofix.core.VideoPlayer import VideoLayer, VideoPlayerError
-from fofix.core.GameEngine import GameEngine
-from fofix.game.MainMenu import MainMenu
-from fofix.core.Language import _
-from fofix.core import Config
-
-class Main():
-    def __init__(self):
-        global args
-        self.args = args
-
-        self.playing = self.args['song']
-        self.configFile = self.args['config']
-        self.fullscreen = self.args['fullscreen']
-        self.resolution = self.args['resolution']
-        self.theme = self.args['theme']
-        self.diff = self.args['diff']
-        self.part = self.args['part']
-        self.mode = self.args['mode']
-        self.players = self.args['players']
-
-        self.config = self.load_config(self.configFile)
-
-        #Lysdestic - Allow support for manipulating fullscreen via CLI
-        if self.fullscreen is not None:
-            Config.set("video", "fullscreen", self.fullscreen)
-
-        #Lysdestic - Change resolution from CLI
-        if self.resolution is not None:
-            Config.set("video", "resolution", self.resolution)
-
-        #Lysdestic - Alter theme from CLI
-        if self.theme is not None:
-            Config.set("coffee", "themename", self.theme)
-
-        self.engine = GameEngine(self.config)
-
-        self.init_oneshot()
-
-        self.videoLayer = False
-        self.restartRequested = False
-
-    @staticmethod
-    def load_config(configPath):
-        ''' Load the configuration file. '''
-        if configPath is not None:
-            if configPath.lower() == "reset":
-
-                # Get os specific location of config file, and remove it.
-                fileName = os.path.join(VFS.getWritableResourcePath(), Version.PROGRAM_UNIXSTYLE_NAME + ".ini")
-                os.remove(fileName)
-
-                # Recreate it
-                config = Config.load(Version.PROGRAM_UNIXSTYLE_NAME + ".ini", setAsDefault = True)
-
-            else:
-                # Load specified config file
-                config = Config.load(configPath, setAsDefault = True)
-        else:
-            # Use default configuration file
-            config = Config.load(Version.PROGRAM_UNIXSTYLE_NAME + ".ini", setAsDefault = True)
-
-        return config
-
-    def init_oneshot(self):
-        ''' Determine if oneshot mode is valid. '''
-        # I think this code can be moved elsewhere...
-        self.engine.cmdPlay = 0
-
-        # Check for a valid invocation of one-shot mode.
-        if self.playing is not None:
-            log.debug('Validating song directory for one-shot mode.')
-
-            library = Config.get("setlist","base_library")
-            basefolder = os.path.join(Version.dataPath(),library,"songs",self.playing)
-
-
-            if not os.path.exists(os.path.join(basefolder, "song.ini")):
-
-                if not (os.path.exists(os.path.join(basefolder, "notes.mid")) or
-                        os.path.exists(os.path.join(basefolder, "notes-unedited.mid"))):
-
-                    if not (os.path.exists(os.path.join(basefolder, "song.ogg")) or
-                            os.path.exists(os.path.join(basefolder, "guitar.ogg"))):
-
-                        log.warn("Song directory provided ('%s') is not a valid song directory. Starting up FoFiX in standard mode." % self.playing)
-                        self.engine.startupMessages.append(_("Song directory provided ('%s') is not a valid song directory. Starting up FoFiX in standard mode.") % self.playing)
-                        return
-
-            # Set up one-shot mode
-            log.debug('Entering one-shot mode.')
-            Config.set("setlist", "selected_song", playing)
-
-            self.engine.cmdPlay = 1
-
-            if diff is not None:
-                self.engine.cmdDiff = int(diff)
-            if part is not None:
-                self.engine.cmdPart = int(part)
-
-            if players == 1:
-                self.engine.cmdMode = players, mode, 0
-            else:
-                self.engine.cmdMode = players, 0, mode
-
-    def restart(self):
-        log.notice("Restarting.")
-        self.engine.audio.close()
-        self.restartRequested = True
-
-    def run(self):
-
-        # Perhapse this could be implemented in a better way...
-        # Play the intro video if it is present, we have the capability, and
-        # we are not in one-shot mode.
-        if not self.engine.cmdPlay:
-            themename = Config.get("coffee", "themename")
-            vidSource = os.path.join(Version.dataPath(), 'themes', themename, 'menu', 'intro.ogv')
-            if os.path.isfile(vidSource):
-                try:
-                    vidPlayer = VideoLayer(self.engine, vidSource, cancellable=True)
-                except (IOError, VideoPlayerError):
-                    log.error("Error loading intro video:")
-                else:
-                    vidPlayer.play()
-                    self.engine.view.pushLayer(vidPlayer)
-                    self.videoLayer = True
-                    self.engine.ticksAtStart = pygame.time.get_ticks()
-                    while not vidPlayer.finished:
-                        self.engine.run()
-                    self.engine.view.popLayer(vidPlayer)
-                    self.engine.view.pushLayer(MainMenu(self.engine))
-        if not self.videoLayer:
-            self.engine.setStartupLayer(MainMenu(self.engine))
-
-        # Run the main game loop.
-        try:
-            self.engine.ticksAtStart = pygame.time.get_ticks()
-            while self.engine.run():
-                pass
-        except KeyboardInterrupt:
-            log.notice("Left mainloop due to KeyboardInterrupt.")
-            # don't reraise
-
-        # Restart the program if the engine is asking that we do so.
-        if self.engine.restartRequested:
-            self.restart()
-
-        # evilynux - MainMenu class already calls this - useless?
-        self.engine.quit()
 
 if __name__ == '__main__':
+    is_windows = (os.name == 'nt')
+    is_macos = sys.platform.startswith('darwin')
+
+    # This prevents the following message being displayed on macOS:
+    # ApplePersistenceIgnoreState: Existing state will not be touched. New state will be written to *path*
+    if is_macos:
+        data = run_command('defaults read org.python.python ApplePersistenceIgnoreState')
+        if data in ['1', 'ON']:
+            run_command('defaults write org.python.python ApplePersistenceIgnoreState 0')
+            atexit.register(run_command, 'defaults write org.python.python ApplePersistenceIgnoreState %s' % data)
+
+    # Add the directory of DLL dependencies to the PATH if we're running
+    # from source on Windows so we pick them up when those bits are imported.
+    if is_windows and not hasattr(sys, 'frozen'):
+        os.environ['PATH'] = os.path.abspath(os.path.join('.', 'win32', 'deps', 'bin')) + os.pathsep + os.environ['PATH']
+
+    args = cmd_args()
+
+    # stump: disable pyOpenGL error checking if we are not asked for it.
+    # This must be before *anything* that may import pyOpenGL!
+    if not args['gl_error_check']:
+        disable_gl_checks()
+
+    # setup the logfile
+    # File object representing the logfile.
+    if is_macos:
+        # evilynux - Under MacOS X, put the logs in ~/Library/Logs
+        logFile = open(os.path.expanduser('~/Library/Logs/%s.log' % Version.PROGRAM_UNIXSTYLE_NAME), 'w')
+    else:
+        logFile = VFS.open('/userdata/%s.log' % Version.PROGRAM_UNIXSTYLE_NAME, 'w')
+
+    log.setLogfile(logFile)
+
+    import fretwork
+    fretworkRequired = (0, 2, 0)
+    reqVerStr = '.'.join([str(i) for i in fretworkRequired])
+    fretworkErrorStr = '''
+
+    The version of fretwork installed is old. Please install the latest version from github.
+    https://github.com/fofix/fretwork/releases/
+    Installed: {0}
+    Required: {1}
+    '''
+    if hasattr(fretwork, '__version__'):  # The first version of fretwork didnt have __version__
+        version = fretwork.__version__.split('-')[0]  # remove 'dev' from ver
+        version = tuple([int(i) for i in version.split('.')])
+
+        if version < fretworkRequired:
+            fwErrStr = fretworkErrorStr.format(fretwork.__version__, reqVerStr)
+            raise RuntimeError(fwErrStr)
+
+    else:
+        version = '0.1.0'
+        fwErrStr = fretworkErrorStr.format(version, reqVerStr)
+        raise RuntimeError(fwErrStr)
+
+    from fofix.game.Main import Main
     try:
         # This loop restarts the game if a restart is requested
         while True:
-            main = Main()
+            main = Main(args)
             main.run()
             if not main.restartRequested:
                 break
@@ -325,15 +168,15 @@ if __name__ == '__main__':
           _logname,
           _("The log file already includes the traceback given above."))
 
-        if os.name == 'nt':
+        if is_windows:
             # If we move to PySDL2 we can replace this with a call to SDL_ShowSimpleMessageBox
             import win32api
             import win32con
-            if win32api.MessageBox(0, "%s\n\n%s" % (_errmsg, _("Open the logfile now?")), "%s %s" % (Version.PROGRAM_NAME, Version.version()), win32con.MB_YESNO|win32con.MB_ICONSTOP) == win32con.IDYES:
+            if win32api.MessageBox(0, "%s\n\n%s" % (_errmsg, _("Open the logfile now?")), "%s %s" % (Version.PROGRAM_NAME, Version.version()), win32con.MB_YESNO | win32con.MB_ICONSTOP) == win32con.IDYES:
                 log.logFile.close()
                 os.startfile(_logname)
             if hasattr(sys, 'frozen'):
                 sys.exit(1)  # don't reraise if py2exe'd so the "Errors occurred" box won't appear after this and confuse the user as to which logfile we actually want
         else:
-            print >>sys.stderr, _errmsg
+            print(_errmsg, file=sys.stderr)
         raise
